Add common exception handling, return concrete types as preference

+++ UKSF.Api.Integrations.Teamspeak/Models/TeamspeakGroupProcedure.cs
@@ -1,6 +1,8 @@
-﻿namespace UKSF.Api.Teamspeak.Models {
-    public class TeamspeakGroupProcedure {
-        public double ClientDbId;
-        public double ServerGroup;
+﻿namespace UKSF.Api.Teamspeak.Models
+{
+    public class TeamspeakGroupProcedure
+    {
+        public int ClientDbId;
+        public int ServerGroup;
     }
 }
